refactor(app): add more namespacing for command text utils
