+++ app/src/local-resources/commands/hooks/useCommandTextString/utils/commandText/getAbsorbanceReaderCommandText.ts
@@ -5,7 +5,7 @@ import type {
   AbsorbanceReaderReadRunTimeCommand,
   RunTimeCommand,
 } from '@opentrons/shared-data'
-import type { HandlesCommands } from './types'
+import type { HandlesCommands } from '../types'
 
 export type AbsorbanceCreateCommand =
   | AbsorbanceReaderOpenLidRunTimeCommand
