+++ app/src/local-resources/commands/hooks/useCommandTextString/utils/commandText/getCommentCommandText.ts
@@ -1,5 +1,5 @@
 import type { CommentRunTimeCommand } from '@opentrons/shared-data/command'
-import type { HandlesCommands } from './types'
+import type { HandlesCommands } from '../types'
 
 export function getCommentCommandText({
   command,
