+++ app/src/local-resources/commands/hooks/useCommandTextString/utils/commandText/getCustomCommandText.ts
@@ -1,5 +1,5 @@
 import type { CustomRunTimeCommand } from '@opentrons/shared-data/command'
-import type { HandlesCommands } from './types'
+import type { HandlesCommands } from '../types'
 
 export function getCustomCommandText({
   command,
