+++ app/src/local-resources/commands/hooks/useCommandTextString/utils/commandText/getRailLightsCommandText.ts
@@ -1,5 +1,5 @@
 import type { RunTimeCommand } from '@opentrons/shared-data/command'
-import type { HandlesCommands } from './types'
+import type { HandlesCommands } from '../types'
 
 type HandledCommands = Extract<RunTimeCommand, { commandType: 'setRailLights' }>
