+++ app/src/local-resources/commands/hooks/useCommandTextString/utils/commandText/getUnknownCommandText.ts
@@ -1,4 +1,4 @@
-import type { GetCommandText } from '..'
+import type { GetCommandText } from '../..'
 
 export function getUnknownCommandText({ command }: GetCommandText): string {
   return JSON.stringify(command)
