+++ app/src/local-resources/commands/hooks/useCommandTextString/utils/commandText/getWaitForDurationCommandText.ts
@@ -1,5 +1,5 @@
 import type { WaitForDurationRunTimeCommand } from '@opentrons/shared-data/command'
-import type { HandlesCommands } from './types'
+import type { HandlesCommands } from '../types'
 
 export function getWaitForDurationCommandText({
   command,
