+++ app/src/local-resources/commands/hooks/useCommandTextString/utils/commandText/getWaitForResumeCommandText.ts
@@ -1,5 +1,5 @@
 import type { WaitForResumeRunTimeCommand } from '@opentrons/shared-data/command'
-import type { HandlesCommands } from './types'
+import type { HandlesCommands } from '../types'
 
 export function getWaitForResumeCommandText({
   command,
