+++ app/src/local-resources/commands/hooks/useCommandTextString/utils/index.ts
@@ -1,26 +1,2 @@
-export { getLoadCommandText } from './getLoadCommandText'
-export { getTemperatureCommandText } from './getTemperatureCommandText'
-export { getTCRunProfileCommandText } from './getTCRunProfileCommandText'
-export { getTCRunExtendedProfileCommandText } from './getTCRunExtendedProfileCommandText'
-export { getHSShakeSpeedCommandText } from './getHSShakeSpeedCommandText'
-export { getMoveToSlotCommandText } from './getMoveToSlotCommandText'
-export { getMoveRelativeCommandText } from './getMoveRelativeCommandText'
-export { getMoveToCoordinatesCommandText } from './getMoveToCoordinatesCommandText'
-export { getMoveToWellCommandText } from './getMoveToWellCommandText'
-export { getMoveLabwareCommandText } from './getMoveLabwareCommandText'
-export { getConfigureForVolumeCommandText } from './getConfigureForVolumeCommandText'
-export { getConfigureNozzleLayoutCommandText } from './getConfigureNozzleLayoutCommandText'
-export { getPrepareToAspirateCommandText } from './getPrepareToAspirateCommandText'
-export { getMoveToAddressableAreaCommandText } from './getMoveToAddressableAreaCommandText'
-export { getMoveToAddressableAreaForDropTipCommandText } from './getMoveToAddressabelAreaForDropTipCommandText'
-export { getDirectTranslationCommandText } from './getDirectTranslationCommandText'
-export { getWaitForDurationCommandText } from './getWaitForDurationCommandText'
-export { getWaitForResumeCommandText } from './getWaitForResumeCommandText'
-export { getDelayCommandText } from './getDelayCommandText'
-export { getCommentCommandText } from './getCommentCommandText'
-export { getCustomCommandText } from './getCustomCommandText'
-export { getUnknownCommandText } from './getUnknownCommandText'
-export { getPipettingCommandText } from './getPipettingCommandText'
-export { getLiquidProbeCommandText } from './getLiquidProbeCommandText'
-export { getRailLightsCommandText } from './getRailLightsCommandText'
-export { getAbsorbanceReaderCommandText } from './getAbsorbanceReaderCommandText'
+export * from './commandText'
+export * from './types'
